requirements.txt
accelerate==0.33.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
base58==2.1.1
bidict==0.23.1
bracex==2.5.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
diffusers @ git+https://github.com/huggingface/diffusers.git@e47cc1fc1a89a5375c322d296cd122fe71ab859f
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
ghp-import==2.1.0
h11==0.14.0
huggingface-hub==0.26.2
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
ipfshttpclient==0.7.0
Jinja2==3.1.4
lazydocs==0.4.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-awesome-pages-plugin==2.9.3
mkdocs-get-deps==0.2.0
mpmath==1.3.0
multiaddr==0.0.9
multidict==6.1.0
natsort==8.4.0
netaddr==1.3.0
networkx==3.4.2
numpy==1.26.4
packaging==24.2
pathspec==0.12.1
payments-py==0.5.2
pillow==11.0.0
pinatapy-vourhey==0.2.0
platformdirs==4.3.6
pluggy==1.5.0
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
pydantic==2.10.2
pydantic_core==2.27.1
Pygments==2.18.0
PyJWT==2.10.0
pytest==8.3.3
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
PyYAML==6.0.2
pyyaml_env_tag==0.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.4.5
scipy==1.14.1
sentencepiece==0.2.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
sympy==1.13.1
tokenizers==0.19.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.43.3
typer==0.13.1
typing_extensions==4.12.2
urllib3==2.2.3
varint==1.0.2
watchdog==6.0.0
wcmatch==10.0
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.18.0
zipp==3.21.0
